move autograd folder from tools to torchgen
[ghstack-poisoned]
