+++ tools/setup_helpers/generate_code.py
@@ -28,8 +28,8 @@ def generate_code(
 ) -> None:
     from torchgen.selective_build.selector import SelectiveBuilder
 
-    from tools.autograd.gen_annotated_fn_args import gen_annotated
-    from tools.autograd.gen_autograd import gen_autograd, gen_autograd_python
+    from torchgen.autograd.gen_annotated_fn_args import gen_annotated
+    from torchgen.autograd.gen_autograd import gen_autograd, gen_autograd_python
 
     # Build ATen based Variable classes
     if install_dir is None:
